scripts/infra-mcp/pyproject.toml
[project]
name = "infra-mcp"
version = "0.1.0"
description = "MCP server for task runner integration, container operations, and infrastructure management"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
dependencies = [
    "fastmcp>=2.10.0,<3",
    "requests>=2.32.4,<3",
    "beautifulsoup4>=4.13,<5",
    "pyyaml>=6.0.2,<7",
]

[project.scripts]
infra-mcp = "server:main"
